Tweak reducer and improve Search input
Created a copy of the voices, being 'filteredVoices', this way it gets isolated from the base list of voices.
Also, add the use case of the search input value to be cleared/reset. This way when the input gets reset (clicked on the X right-hand side) it's displaying the full list of voices.